requirements.txt
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
flower==2.0.1
humanize==4.10.0
kombu==5.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
six==1.16.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
pyyaml==6.0.2

# lint and format
black==24.8.0
flake8==7.1.1
isort==5.13.2
